Adding code to turn it into a terminal
 - Reads from serial and writes the data to stdout
 - Reads from stdin and writes the data to serial

A problem is after writing a CR (aka. 13 decimal) to
the serial at the end of the data from stdin. The bytes_to_read
continuously returns a "I/O error"!

Also, the ee203 is coming up in "Debug mode" and prompts
with "M>" instead of the expected "CMD>".

When using minicom I also see a "DBG>" prompt and after typing "h" CR
on minicom it flashes a dialog to the termainl for about 1 second saying
"Cannot open /dev/ttyACM0".